move shared methods into HearingConcern
