+++ app/models/concerns/hearing_concern.rb
@@ -1,0 +1,28 @@
+# frozen_string_literal: true
+
+##
+# Shared methods of Hearing and LegacyHearing models
+##
+module HearingConcern
+  extend ActiveSupport::Concern
+
+  # NOTE: for LegacyHearing, this makes a call to VACOLS
+  def postponed?
+    disposition == Constants.HEARING_DISPOSITION_TYPES.postponed
+  end
+
+  # NOTE: for LegacyHearing, this makes a call to VACOLS
+  def cancelled?
+    disposition == Constants.HEARING_DISPOSITION_TYPES.cancelled
+  end
+
+  # NOTE: for LegacyHearing, this makes a call to VACOLS
+  def no_show?
+    disposition == Constants.HEARING_DISPOSITION_TYPES.no_show
+  end
+
+  # NOTE: for LegacyHearing, this makes a call to VACOLS
+  def held?
+    disposition == Constants.HEARING_DISPOSITION_TYPES.held
+  end
+end
